+++ tests-bls/tests/harness.rs
@@ -1,4 +1,5 @@
 mod tests_add_fp;
 mod tests_mul_by_x_fp;
 mod tests_sub_fp;
-mod tests_temp_fe_mont_mul;
+mod tests_l_shift;
+// mod tests_temp_fe_mont_mul;
